arch/risc-v/src/mpfs/mpfs_dsn: Correct serial number reading routine
- Fix retry counter handling and increase the maximum number of retries.
- Add some parenthesis for clarity.
- Change return type to signed int, as it may return -ETIMEDOUT
- Correct comment for returned value

+++ arch/risc-v/src/mpfs/mpfs_dsn.h
@@ -70,7 +70,7 @@ extern "C"
  *
  ****************************************************************************/
 
-size_t mpfs_read_dsn(uint8_t *dsn, size_t len);
+int mpfs_read_dsn(uint8_t *dsn, size_t len);
 
 #ifdef __cplusplus
 }
